requirements.txt
tensorflow>=2.15
numpy>=1.24
Pillow>=9.5